Revert "dev/settings"
