+++ lib/index.ts
@@ -1,2 +1,1 @@
 export { device } from './device'
-export { theme } from './theme'
