--- lib/theme.ts
@@ -1,10 +1,0 @@
-const theme = {
-  colors: {
-    primary: '#57cc8a',
-    gray: '#757575',
-    bgDark: '#242930',
-    bgDarkGray: 'rgba(175, 186, 196, .25)'
-  },
-} as const
-
-export { theme }
