+++ package.json
@@ -5,8 +5,7 @@
   "scripts": {
     "dev": "NODE_OPTIONS='--inspect' next dev",
     "build": "next build",
-    "start": "next start",
-    "lint": "eslint components/**/*.{ts,tsx} lib/**/*.ts pages/*.tsx"
+    "start": "next start"
   },
   "dependencies": {
     "github-markdown-css": "^4.0.0",
@@ -24,13 +23,7 @@
     "@types/react": "^16.9.56",
     "@types/react-syntax-highlighter": "^13.5.0",
     "@types/styled-components": "^5.1.4",
-    "@typescript-eslint/eslint-plugin": "^4.8.1",
-    "@typescript-eslint/parser": "^4.8.1",
     "babel-plugin-styled-components": "^1.11.1",
-    "eslint": "^7.14.0",
-    "eslint-config-prettier": "^6.15.0",
-    "eslint-plugin-react": "^7.21.5",
-    "prettier": "2.2.0",
     "typescript": "^4.0.5"
   }
 }
