+++ pages/index.tsx
@@ -5,7 +5,7 @@ import { BlogList, Layout } from '../components'
 
 type Props = { data: string[]; title: string; description: string }
 
-const MyApp: React.FC<Props> = ({ data, title, description }) => {
+const MyApp = ({ data, title, description }: Props) => {
   const blogs = data.map(blog => matter(blog)).map(item => item.data)
 
   return (
@@ -29,7 +29,7 @@ const MyApp: React.FC<Props> = ({ data, title, description }) => {
 
 export const getStaticProps: GetStaticProps = async () => {
   const siteData = await import("../config.json")
-  const fs = await import("fs")
+  const fs = require("fs")
 
   const files = fs.readdirSync(`${process.cwd()}/content`, "utf-8")
   const blogs = files.filter(fn => fn.endsWith(".md"))
