+++ tsconfig.json
@@ -12,8 +12,7 @@
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "jsx": "preserve",
-    "typeRoots": ["types", "node_modules/@types"]
+    "jsx": "preserve"
   },
   "exclude": [ "node_modules" ],
   "include": [ "next-env.d.ts", "**/*.ts", "**/*.tsx" ]
